Move screen space effects into a separate class
