+++ servers/rendering/renderer_rd/effects_rd.cpp
@@ -41,14 +41,6 @@ bool EffectsRD::get_prefer_raster_effects() {
 	return prefer_raster_effects;
 }
 
-static _FORCE_INLINE_ void store_camera(const CameraMatrix &p_mtx, float *p_array) {
-	for (int i = 0; i < 4; i++) {
-		for (int j = 0; j < 4; j++) {
-			p_array[i * 4 + j] = p_mtx.matrix[i][j];
-		}
-	}
-}
-
 RID EffectsRD::_get_uniform_set_from_image(RID p_image) {
 	if (image_to_uniform_set_cache.has(p_image)) {
 		RID uniform_set = image_to_uniform_set_cache[p_image];
@@ -86,7 +78,7 @@ RID EffectsRD::_get_uniform_set_from_texture(RID p_texture, bool p_use_mipmaps) 
 	u.append_id(p_texture);
 	uniforms.push_back(u);
 	// anything with the same configuration (one texture in binding 0 for set 0), is good
-	RID uniform_set = RD::get_singleton()->uniform_set_create(uniforms, specular_merge.shader.version_get_shader(specular_merge.shader_version, 0), 0);
+	RID uniform_set = RD::get_singleton()->uniform_set_create(uniforms, luminance_reduce_raster.shader.version_get_shader(luminance_reduce_raster.shader_version, 0), 0);
 
 	texture_to_uniform_set_cache[p_texture] = uniform_set;
 
@@ -112,105 +104,6 @@ RID EffectsRD::_get_compute_uniform_set_from_texture(RID p_texture, bool p_use_m
 	RID uniform_set = RD::get_singleton()->uniform_set_create(uniforms, luminance_reduce.shader.version_get_shader(luminance_reduce.shader_version, 0), 0);
 
 	texture_to_compute_uniform_set_cache[p_texture] = uniform_set;
-
-	return uniform_set;
-}
-
-RID EffectsRD::_get_compute_uniform_set_from_texture_and_sampler(RID p_texture, RID p_sampler) {
-	TextureSamplerPair tsp;
-	tsp.texture = p_texture;
-	tsp.sampler = p_sampler;
-
-	if (texture_sampler_to_compute_uniform_set_cache.has(tsp)) {
-		RID uniform_set = texture_sampler_to_compute_uniform_set_cache[tsp];
-		if (RD::get_singleton()->uniform_set_is_valid(uniform_set)) {
-			return uniform_set;
-		}
-	}
-
-	Vector<RD::Uniform> uniforms;
-	RD::Uniform u;
-	u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-	u.binding = 0;
-	u.append_id(p_sampler);
-	u.append_id(p_texture);
-	uniforms.push_back(u);
-	//any thing with the same configuration (one texture in binding 0 for set 0), is good
-	RID uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssao.blur_shader.version_get_shader(ssao.blur_shader_version, 0), 0);
-
-	texture_sampler_to_compute_uniform_set_cache[tsp] = uniform_set;
-
-	return uniform_set;
-}
-
-RID EffectsRD::_get_compute_uniform_set_from_texture_pair(RID p_texture1, RID p_texture2, bool p_use_mipmaps) {
-	TexturePair tp;
-	tp.texture1 = p_texture1;
-	tp.texture2 = p_texture2;
-
-	if (texture_pair_to_compute_uniform_set_cache.has(tp)) {
-		RID uniform_set = texture_pair_to_compute_uniform_set_cache[tp];
-		if (RD::get_singleton()->uniform_set_is_valid(uniform_set)) {
-			return uniform_set;
-		}
-	}
-
-	Vector<RD::Uniform> uniforms;
-	{
-		RD::Uniform u;
-		u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-		u.binding = 0;
-		u.append_id(p_use_mipmaps ? default_mipmap_sampler : default_sampler);
-		u.append_id(p_texture1);
-		uniforms.push_back(u);
-	}
-	{
-		RD::Uniform u;
-		u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-		u.binding = 1;
-		u.append_id(p_use_mipmaps ? default_mipmap_sampler : default_sampler);
-		u.append_id(p_texture2);
-		uniforms.push_back(u);
-	}
-	//any thing with the same configuration (one texture in binding 0 for set 0), is good
-	RID uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssr_scale.shader.version_get_shader(ssr_scale.shader_version, 0), 1);
-
-	texture_pair_to_compute_uniform_set_cache[tp] = uniform_set;
-
-	return uniform_set;
-}
-
-RID EffectsRD::_get_compute_uniform_set_from_image_pair(RID p_texture1, RID p_texture2) {
-	TexturePair tp;
-	tp.texture1 = p_texture1;
-	tp.texture2 = p_texture2;
-
-	if (image_pair_to_compute_uniform_set_cache.has(tp)) {
-		RID uniform_set = image_pair_to_compute_uniform_set_cache[tp];
-		if (RD::get_singleton()->uniform_set_is_valid(uniform_set)) {
-			return uniform_set;
-		}
-	}
-
-	Vector<RD::Uniform> uniforms;
-	{
-		RD::Uniform u;
-		u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-		u.binding = 0;
-		u.append_id(p_texture1);
-		uniforms.push_back(u);
-	}
-	{
-		RD::Uniform u;
-		u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-		u.binding = 1;
-		u.append_id(p_texture2);
-		uniforms.push_back(u);
-	}
-	//any thing with the same configuration (one texture in binding 0 for set 0), is good
-	RID uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssr_scale.shader.version_get_shader(ssr_scale.shader_version, 0), 3);
-
-	image_pair_to_compute_uniform_set_cache[tp] = uniform_set;
 
 	return uniform_set;
 }
@@ -281,125 +174,6 @@ void EffectsRD::taa_resolve(RID p_frame, RID p_temp, RID p_depth, RID p_velocity
 	RD::get_singleton()->compute_list_end();
 }
 
-void EffectsRD::screen_space_reflection(RID p_diffuse, RID p_normal_roughness, RenderingServer::EnvironmentSSRRoughnessQuality p_roughness_quality, RID p_blur_radius, RID p_blur_radius2, RID p_metallic, const Color &p_metallic_mask, RID p_depth, RID p_scale_depth, RID p_scale_normal, RID p_output, RID p_output_blur, const Size2i &p_screen_size, int p_max_steps, float p_fade_in, float p_fade_out, float p_tolerance, const CameraMatrix &p_camera) {
-	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-
-	{ //scale color and depth to half
-		ssr_scale.push_constant.camera_z_far = p_camera.get_z_far();
-		ssr_scale.push_constant.camera_z_near = p_camera.get_z_near();
-		ssr_scale.push_constant.orthogonal = p_camera.is_orthogonal();
-		ssr_scale.push_constant.filter = false; //enabling causes arctifacts
-		ssr_scale.push_constant.screen_size[0] = p_screen_size.x;
-		ssr_scale.push_constant.screen_size[1] = p_screen_size.y;
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssr_scale.pipeline);
-
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_diffuse), 0);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture_pair(p_depth, p_normal_roughness), 1);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_output_blur), 2);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_image_pair(p_scale_depth, p_scale_normal), 3);
-
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssr_scale.push_constant, sizeof(ScreenSpaceReflectionScalePushConstant));
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_screen_size.width, p_screen_size.height, 1);
-
-		RD::get_singleton()->compute_list_add_barrier(compute_list);
-	}
-
-	{
-		ssr.push_constant.camera_z_far = p_camera.get_z_far();
-		ssr.push_constant.camera_z_near = p_camera.get_z_near();
-		ssr.push_constant.orthogonal = p_camera.is_orthogonal();
-		ssr.push_constant.screen_size[0] = p_screen_size.x;
-		ssr.push_constant.screen_size[1] = p_screen_size.y;
-		ssr.push_constant.curve_fade_in = p_fade_in;
-		ssr.push_constant.distance_fade = p_fade_out;
-		ssr.push_constant.num_steps = p_max_steps;
-		ssr.push_constant.depth_tolerance = p_tolerance;
-		ssr.push_constant.use_half_res = true;
-		ssr.push_constant.proj_info[0] = -2.0f / (p_screen_size.width * p_camera.matrix[0][0]);
-		ssr.push_constant.proj_info[1] = -2.0f / (p_screen_size.height * p_camera.matrix[1][1]);
-		ssr.push_constant.proj_info[2] = (1.0f - p_camera.matrix[0][2]) / p_camera.matrix[0][0];
-		ssr.push_constant.proj_info[3] = (1.0f + p_camera.matrix[1][2]) / p_camera.matrix[1][1];
-		ssr.push_constant.metallic_mask[0] = CLAMP(p_metallic_mask.r * 255.0, 0, 255);
-		ssr.push_constant.metallic_mask[1] = CLAMP(p_metallic_mask.g * 255.0, 0, 255);
-		ssr.push_constant.metallic_mask[2] = CLAMP(p_metallic_mask.b * 255.0, 0, 255);
-		ssr.push_constant.metallic_mask[3] = CLAMP(p_metallic_mask.a * 255.0, 0, 255);
-		store_camera(p_camera, ssr.push_constant.projection);
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssr.pipelines[(p_roughness_quality != RS::ENV_SSR_ROUGHNESS_QUALITY_DISABLED) ? SCREEN_SPACE_REFLECTION_ROUGH : SCREEN_SPACE_REFLECTION_NORMAL]);
-
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssr.push_constant, sizeof(ScreenSpaceReflectionPushConstant));
-
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_image_pair(p_output_blur, p_scale_depth), 0);
-
-		if (p_roughness_quality != RS::ENV_SSR_ROUGHNESS_QUALITY_DISABLED) {
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_image_pair(p_output, p_blur_radius), 1);
-		} else {
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_output), 1);
-		}
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_metallic), 3);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_scale_normal), 2);
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_screen_size.width, p_screen_size.height, 1);
-	}
-
-	if (p_roughness_quality != RS::ENV_SSR_ROUGHNESS_QUALITY_DISABLED) {
-		//blur
-
-		RD::get_singleton()->compute_list_add_barrier(compute_list);
-
-		ssr_filter.push_constant.orthogonal = p_camera.is_orthogonal();
-		ssr_filter.push_constant.edge_tolerance = Math::sin(Math::deg2rad(15.0));
-		ssr_filter.push_constant.proj_info[0] = -2.0f / (p_screen_size.width * p_camera.matrix[0][0]);
-		ssr_filter.push_constant.proj_info[1] = -2.0f / (p_screen_size.height * p_camera.matrix[1][1]);
-		ssr_filter.push_constant.proj_info[2] = (1.0f - p_camera.matrix[0][2]) / p_camera.matrix[0][0];
-		ssr_filter.push_constant.proj_info[3] = (1.0f + p_camera.matrix[1][2]) / p_camera.matrix[1][1];
-		ssr_filter.push_constant.vertical = 0;
-		if (p_roughness_quality == RS::ENV_SSR_ROUGHNESS_QUALITY_LOW) {
-			ssr_filter.push_constant.steps = p_max_steps / 3;
-			ssr_filter.push_constant.increment = 3;
-		} else if (p_roughness_quality == RS::ENV_SSR_ROUGHNESS_QUALITY_MEDIUM) {
-			ssr_filter.push_constant.steps = p_max_steps / 2;
-			ssr_filter.push_constant.increment = 2;
-		} else {
-			ssr_filter.push_constant.steps = p_max_steps;
-			ssr_filter.push_constant.increment = 1;
-		}
-
-		ssr_filter.push_constant.screen_size[0] = p_screen_size.width;
-		ssr_filter.push_constant.screen_size[1] = p_screen_size.height;
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssr_filter.pipelines[SCREEN_SPACE_REFLECTION_FILTER_HORIZONTAL]);
-
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_image_pair(p_output, p_blur_radius), 0);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_scale_normal), 1);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_image_pair(p_output_blur, p_blur_radius2), 2);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_scale_depth), 3);
-
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssr_filter.push_constant, sizeof(ScreenSpaceReflectionFilterPushConstant));
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_screen_size.width, p_screen_size.height, 1);
-
-		RD::get_singleton()->compute_list_add_barrier(compute_list);
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssr_filter.pipelines[SCREEN_SPACE_REFLECTION_FILTER_VERTICAL]);
-
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_image_pair(p_output_blur, p_blur_radius2), 0);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_scale_normal), 1);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_output), 2);
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_scale_depth), 3);
-
-		ssr_filter.push_constant.vertical = 1;
-
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssr_filter.push_constant, sizeof(ScreenSpaceReflectionFilterPushConstant));
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_screen_size.width, p_screen_size.height, 1);
-	}
-
-	RD::get_singleton()->compute_list_end();
-}
-
 void EffectsRD::sub_surface_scattering(RID p_diffuse, RID p_diffuse2, RID p_depth, const CameraMatrix &p_camera, const Size2i &p_screen_size, float p_scale, float p_depth_scale, RenderingServer::SubSurfaceScatteringQuality p_quality) {
 	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
 
@@ -441,36 +215,6 @@ void EffectsRD::sub_surface_scattering(RID p_diffuse, RID p_diffuse2, RID p_dept
 
 		RD::get_singleton()->compute_list_end();
 	}
-}
-
-void EffectsRD::merge_specular(RID p_dest_framebuffer, RID p_specular, RID p_base, RID p_reflection) {
-	RD::DrawListID draw_list = RD::get_singleton()->draw_list_begin(p_dest_framebuffer, RD::INITIAL_ACTION_KEEP, RD::FINAL_ACTION_READ, RD::INITIAL_ACTION_KEEP, RD::FINAL_ACTION_READ, Vector<Color>());
-
-	if (p_reflection.is_valid()) {
-		if (p_base.is_valid()) {
-			RD::get_singleton()->draw_list_bind_render_pipeline(draw_list, specular_merge.pipelines[SPECULAR_MERGE_SSR].get_render_pipeline(RD::INVALID_ID, RD::get_singleton()->framebuffer_get_format(p_dest_framebuffer)));
-			RD::get_singleton()->draw_list_bind_uniform_set(draw_list, _get_uniform_set_from_texture(p_base), 2);
-		} else {
-			RD::get_singleton()->draw_list_bind_render_pipeline(draw_list, specular_merge.pipelines[SPECULAR_MERGE_ADDITIVE_SSR].get_render_pipeline(RD::INVALID_ID, RD::get_singleton()->framebuffer_get_format(p_dest_framebuffer)));
-		}
-
-		RD::get_singleton()->draw_list_bind_uniform_set(draw_list, _get_uniform_set_from_texture(p_specular), 0);
-		RD::get_singleton()->draw_list_bind_uniform_set(draw_list, _get_uniform_set_from_texture(p_reflection), 1);
-
-	} else {
-		if (p_base.is_valid()) {
-			RD::get_singleton()->draw_list_bind_render_pipeline(draw_list, specular_merge.pipelines[SPECULAR_MERGE_ADD].get_render_pipeline(RD::INVALID_ID, RD::get_singleton()->framebuffer_get_format(p_dest_framebuffer)));
-			RD::get_singleton()->draw_list_bind_uniform_set(draw_list, _get_uniform_set_from_texture(p_base), 2);
-		} else {
-			RD::get_singleton()->draw_list_bind_render_pipeline(draw_list, specular_merge.pipelines[SPECULAR_MERGE_ADDITIVE_ADD].get_render_pipeline(RD::INVALID_ID, RD::get_singleton()->framebuffer_get_format(p_dest_framebuffer)));
-		}
-
-		RD::get_singleton()->draw_list_bind_uniform_set(draw_list, _get_uniform_set_from_texture(p_specular), 0);
-	}
-
-	RD::get_singleton()->draw_list_bind_index_array(draw_list, index_array);
-	RD::get_singleton()->draw_list_draw(draw_list, true);
-	RD::get_singleton()->draw_list_end();
 }
 
 void EffectsRD::luminance_reduction(RID p_source_texture, const Size2i p_source_size, const Vector<RID> p_reduce, RID p_prev_luminance, float p_min_luminance, float p_max_luminance, float p_adjust, bool p_set) {
@@ -544,674 +288,6 @@ void EffectsRD::luminance_reduction_raster(RID p_source_texture, const Size2i p_
 		RD::get_singleton()->draw_list_draw(draw_list, true);
 		RD::get_singleton()->draw_list_end();
 	}
-}
-
-void EffectsRD::downsample_depth(RID p_depth_buffer, const Vector<RID> &p_depth_mipmaps, RS::EnvironmentSSAOQuality p_ssao_quality, RS::EnvironmentSSILQuality p_ssil_quality, bool p_invalidate_uniform_set, bool p_ssao_half_size, bool p_ssil_half_size, Size2i p_full_screen_size, const CameraMatrix &p_projection) {
-	// Downsample and deinterleave the depth buffer for SSAO and SSIL
-	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-
-	int downsample_pipeline = SS_EFFECTS_DOWNSAMPLE;
-	bool use_mips = p_ssao_quality > RS::ENV_SSAO_QUALITY_MEDIUM || p_ssil_quality > RS::ENV_SSIL_QUALITY_MEDIUM;
-
-	if (p_ssao_quality == RS::ENV_SSAO_QUALITY_VERY_LOW && p_ssil_quality == RS::ENV_SSIL_QUALITY_VERY_LOW) {
-		downsample_pipeline = SS_EFFECTS_DOWNSAMPLE_HALF;
-	} else if (use_mips) {
-		downsample_pipeline = SS_EFFECTS_DOWNSAMPLE_MIPMAP;
-	}
-
-	bool use_half_size = false;
-	bool use_full_mips = false;
-
-	if (p_ssao_half_size && p_ssil_half_size) {
-		downsample_pipeline++;
-		use_half_size = true;
-	} else if (p_ssao_half_size != p_ssil_half_size) {
-		if (use_mips) {
-			downsample_pipeline = SS_EFFECTS_DOWNSAMPLE_FULL_MIPS;
-			use_full_mips = true;
-		} else {
-			// Only need the first two mipmaps, but the cost to generate the next two is trivial
-			// TODO investigate the benefit of a shader version to generate only 2 mips
-			downsample_pipeline = SS_EFFECTS_DOWNSAMPLE_MIPMAP;
-			use_mips = true;
-		}
-	}
-
-	int depth_index = use_half_size ? 1 : 0;
-
-	RD::get_singleton()->draw_command_begin_label("Downsample Depth");
-	if (p_invalidate_uniform_set || use_full_mips != ss_effects.used_full_mips_last_frame || use_half_size != ss_effects.used_half_size_last_frame || use_mips != ss_effects.used_mips_last_frame) {
-		Vector<RD::Uniform> uniforms;
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-			u.binding = 0;
-			u.append_id(p_depth_mipmaps[depth_index + 1]);
-			uniforms.push_back(u);
-		}
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-			u.binding = 1;
-			u.append_id(p_depth_mipmaps[depth_index + 2]);
-			uniforms.push_back(u);
-		}
-		{
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-			u.binding = 2;
-			u.append_id(p_depth_mipmaps[depth_index + 3]);
-			uniforms.push_back(u);
-		}
-		if (use_full_mips) {
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-			u.binding = 3;
-			u.append_id(p_depth_mipmaps[4]);
-			uniforms.push_back(u);
-		}
-		ss_effects.downsample_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ss_effects.downsample_shader.version_get_shader(ss_effects.downsample_shader_version, use_full_mips ? 6 : 2), 2);
-	}
-
-	float depth_linearize_mul = -p_projection.matrix[3][2];
-	float depth_linearize_add = p_projection.matrix[2][2];
-	if (depth_linearize_mul * depth_linearize_add < 0) {
-		depth_linearize_add = -depth_linearize_add;
-	}
-
-	ss_effects.downsample_push_constant.orthogonal = p_projection.is_orthogonal();
-	ss_effects.downsample_push_constant.z_near = depth_linearize_mul;
-	ss_effects.downsample_push_constant.z_far = depth_linearize_add;
-	if (ss_effects.downsample_push_constant.orthogonal) {
-		ss_effects.downsample_push_constant.z_near = p_projection.get_z_near();
-		ss_effects.downsample_push_constant.z_far = p_projection.get_z_far();
-	}
-	ss_effects.downsample_push_constant.pixel_size[0] = 1.0 / p_full_screen_size.x;
-	ss_effects.downsample_push_constant.pixel_size[1] = 1.0 / p_full_screen_size.y;
-	ss_effects.downsample_push_constant.radius_sq = 1.0;
-
-	RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ss_effects.pipelines[downsample_pipeline]);
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_depth_buffer), 0);
-	RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_depth_mipmaps[depth_index + 0]), 1);
-	if (use_mips) {
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, ss_effects.downsample_uniform_set, 2);
-	}
-	RD::get_singleton()->compute_list_set_push_constant(compute_list, &ss_effects.downsample_push_constant, sizeof(SSEffectsDownsamplePushConstant));
-
-	Size2i size(MAX(1, p_full_screen_size.x >> (use_half_size ? 2 : 1)), MAX(1, p_full_screen_size.y >> (use_half_size ? 2 : 1)));
-
-	RD::get_singleton()->compute_list_dispatch_threads(compute_list, size.x, size.y, 1);
-	RD::get_singleton()->compute_list_add_barrier(compute_list);
-	RD::get_singleton()->draw_command_end_label();
-
-	RD::get_singleton()->compute_list_end(RD::BARRIER_MASK_COMPUTE);
-
-	ss_effects.used_full_mips_last_frame = use_full_mips;
-	ss_effects.used_half_size_last_frame = use_half_size;
-}
-
-void EffectsRD::gather_ssao(RD::ComputeListID p_compute_list, const Vector<RID> p_ao_slices, const SSAOSettings &p_settings, bool p_adaptive_base_pass, RID p_gather_uniform_set, RID p_importance_map_uniform_set) {
-	RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, p_gather_uniform_set, 0);
-	if ((p_settings.quality == RS::ENV_SSAO_QUALITY_ULTRA) && !p_adaptive_base_pass) {
-		RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, p_importance_map_uniform_set, 1);
-	}
-
-	for (int i = 0; i < 4; i++) {
-		if ((p_settings.quality == RS::ENV_SSAO_QUALITY_VERY_LOW) && ((i == 1) || (i == 2))) {
-			continue;
-		}
-
-		ssao.gather_push_constant.pass_coord_offset[0] = i % 2;
-		ssao.gather_push_constant.pass_coord_offset[1] = i / 2;
-		ssao.gather_push_constant.pass_uv_offset[0] = ((i % 2) - 0.0) / p_settings.full_screen_size.x;
-		ssao.gather_push_constant.pass_uv_offset[1] = ((i / 2) - 0.0) / p_settings.full_screen_size.y;
-		ssao.gather_push_constant.pass = i;
-		RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, _get_uniform_set_from_image(p_ao_slices[i]), 2);
-		RD::get_singleton()->compute_list_set_push_constant(p_compute_list, &ssao.gather_push_constant, sizeof(SSAOGatherPushConstant));
-
-		Size2i size = Size2i(p_settings.full_screen_size.x >> (p_settings.half_size ? 2 : 1), p_settings.full_screen_size.y >> (p_settings.half_size ? 2 : 1));
-
-		RD::get_singleton()->compute_list_dispatch_threads(p_compute_list, size.x, size.y, 1);
-	}
-	RD::get_singleton()->compute_list_add_barrier(p_compute_list);
-}
-
-void EffectsRD::generate_ssao(RID p_normal_buffer, RID p_depth_mipmaps_texture, RID p_ao, const Vector<RID> p_ao_slices, RID p_ao_pong, const Vector<RID> p_ao_pong_slices, RID p_upscale_buffer, RID p_importance_map, RID p_importance_map_pong, const CameraMatrix &p_projection, const SSAOSettings &p_settings, bool p_invalidate_uniform_sets, RID &r_gather_uniform_set, RID &r_importance_map_uniform_set) {
-	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-	memset(&ssao.gather_push_constant, 0, sizeof(SSAOGatherPushConstant));
-	/* FIRST PASS */
-
-	RD::get_singleton()->draw_command_begin_label("Process Screen Space Ambient Occlusion");
-	/* SECOND PASS */
-	// Sample SSAO
-	{
-		RD::get_singleton()->draw_command_begin_label("Gather Samples");
-		ssao.gather_push_constant.screen_size[0] = p_settings.full_screen_size.x;
-		ssao.gather_push_constant.screen_size[1] = p_settings.full_screen_size.y;
-
-		ssao.gather_push_constant.half_screen_pixel_size[0] = 1.0 / p_settings.half_screen_size.x;
-		ssao.gather_push_constant.half_screen_pixel_size[1] = 1.0 / p_settings.half_screen_size.y;
-		float tan_half_fov_x = 1.0 / p_projection.matrix[0][0];
-		float tan_half_fov_y = 1.0 / p_projection.matrix[1][1];
-		ssao.gather_push_constant.NDC_to_view_mul[0] = tan_half_fov_x * 2.0;
-		ssao.gather_push_constant.NDC_to_view_mul[1] = tan_half_fov_y * -2.0;
-		ssao.gather_push_constant.NDC_to_view_add[0] = tan_half_fov_x * -1.0;
-		ssao.gather_push_constant.NDC_to_view_add[1] = tan_half_fov_y;
-		ssao.gather_push_constant.is_orthogonal = p_projection.is_orthogonal();
-
-		ssao.gather_push_constant.half_screen_pixel_size_x025[0] = ssao.gather_push_constant.half_screen_pixel_size[0] * 0.25;
-		ssao.gather_push_constant.half_screen_pixel_size_x025[1] = ssao.gather_push_constant.half_screen_pixel_size[1] * 0.25;
-
-		ssao.gather_push_constant.radius = p_settings.radius;
-		float radius_near_limit = (p_settings.radius * 1.2f);
-		if (p_settings.quality <= RS::ENV_SSAO_QUALITY_LOW) {
-			radius_near_limit *= 1.50f;
-
-			if (p_settings.quality == RS::ENV_SSAO_QUALITY_VERY_LOW) {
-				ssao.gather_push_constant.radius *= 0.8f;
-			}
-		}
-		radius_near_limit /= tan_half_fov_y;
-		ssao.gather_push_constant.intensity = p_settings.intensity;
-		ssao.gather_push_constant.shadow_power = p_settings.power;
-		ssao.gather_push_constant.shadow_clamp = 0.98;
-		ssao.gather_push_constant.fade_out_mul = -1.0 / (p_settings.fadeout_to - p_settings.fadeout_from);
-		ssao.gather_push_constant.fade_out_add = p_settings.fadeout_from / (p_settings.fadeout_to - p_settings.fadeout_from) + 1.0;
-		ssao.gather_push_constant.horizon_angle_threshold = p_settings.horizon;
-		ssao.gather_push_constant.inv_radius_near_limit = 1.0f / radius_near_limit;
-		ssao.gather_push_constant.neg_inv_radius = -1.0 / ssao.gather_push_constant.radius;
-
-		ssao.gather_push_constant.load_counter_avg_div = 9.0 / float((p_settings.quarter_screen_size.x) * (p_settings.quarter_screen_size.y) * 255);
-		ssao.gather_push_constant.adaptive_sample_limit = p_settings.adaptive_target;
-
-		ssao.gather_push_constant.detail_intensity = p_settings.detail;
-		ssao.gather_push_constant.quality = MAX(0, p_settings.quality - 1);
-		ssao.gather_push_constant.size_multiplier = p_settings.half_size ? 2 : 1;
-
-		if (p_invalidate_uniform_sets) {
-			Vector<RD::Uniform> uniforms;
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-				u.binding = 0;
-				u.append_id(default_sampler);
-				u.append_id(p_depth_mipmaps_texture);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-				u.binding = 1;
-				u.append_id(p_normal_buffer);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_UNIFORM_BUFFER;
-				u.binding = 2;
-				u.append_id(ss_effects.gather_constants_buffer);
-				uniforms.push_back(u);
-			}
-			r_gather_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssao.gather_shader.version_get_shader(ssao.gather_shader_version, 0), 0);
-		}
-
-		if (p_invalidate_uniform_sets) {
-			Vector<RD::Uniform> uniforms;
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-				u.binding = 0;
-				u.append_id(p_ao_pong);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-				u.binding = 1;
-				u.append_id(default_sampler);
-				u.append_id(p_importance_map);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-				u.binding = 2;
-				u.append_id(ssao.importance_map_load_counter);
-				uniforms.push_back(u);
-			}
-			r_importance_map_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssao.gather_shader.version_get_shader(ssao.gather_shader_version, 2), 1);
-		}
-
-		if (p_settings.quality == RS::ENV_SSAO_QUALITY_ULTRA) {
-			RD::get_singleton()->draw_command_begin_label("Generate Importance Map");
-			ssao.importance_map_push_constant.half_screen_pixel_size[0] = 1.0 / p_settings.half_screen_size.x;
-			ssao.importance_map_push_constant.half_screen_pixel_size[1] = 1.0 / p_settings.half_screen_size.y;
-			ssao.importance_map_push_constant.intensity = p_settings.intensity;
-			ssao.importance_map_push_constant.power = p_settings.power;
-			//base pass
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[SSAO_GATHER_BASE]);
-			gather_ssao(compute_list, p_ao_pong_slices, p_settings, true, r_gather_uniform_set, RID());
-			//generate importance map
-
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[SSAO_GENERATE_IMPORTANCE_MAP]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ao_pong), 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_importance_map), 1);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssao.importance_map_push_constant, sizeof(SSAOImportanceMapPushConstant));
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.quarter_screen_size.x, p_settings.quarter_screen_size.y, 1);
-			RD::get_singleton()->compute_list_add_barrier(compute_list);
-			//process importance map A
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[SSAO_PROCESS_IMPORTANCE_MAPA]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_importance_map), 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_importance_map_pong), 1);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssao.importance_map_push_constant, sizeof(SSAOImportanceMapPushConstant));
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.quarter_screen_size.x, p_settings.quarter_screen_size.y, 1);
-			RD::get_singleton()->compute_list_add_barrier(compute_list);
-			//process Importance Map B
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[SSAO_PROCESS_IMPORTANCE_MAPB]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_importance_map_pong), 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_importance_map), 1);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, ssao.counter_uniform_set, 2);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssao.importance_map_push_constant, sizeof(SSAOImportanceMapPushConstant));
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.quarter_screen_size.x, p_settings.quarter_screen_size.y, 1);
-			RD::get_singleton()->compute_list_add_barrier(compute_list);
-
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[SSAO_GATHER_ADAPTIVE]);
-			RD::get_singleton()->draw_command_end_label(); // Importance Map
-		} else {
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[SSAO_GATHER]);
-		}
-
-		gather_ssao(compute_list, p_ao_slices, p_settings, false, r_gather_uniform_set, r_importance_map_uniform_set);
-		RD::get_singleton()->draw_command_end_label(); // Gather SSAO
-	}
-
-	//	/* THIRD PASS */
-	//	// Blur
-	//
-	{
-		RD::get_singleton()->draw_command_begin_label("Edge Aware Blur");
-		ssao.blur_push_constant.edge_sharpness = 1.0 - p_settings.sharpness;
-		ssao.blur_push_constant.half_screen_pixel_size[0] = 1.0 / p_settings.half_screen_size.x;
-		ssao.blur_push_constant.half_screen_pixel_size[1] = 1.0 / p_settings.half_screen_size.y;
-
-		int blur_passes = p_settings.quality > RS::ENV_SSAO_QUALITY_VERY_LOW ? p_settings.blur_passes : 1;
-
-		for (int pass = 0; pass < blur_passes; pass++) {
-			int blur_pipeline = SSAO_BLUR_PASS;
-			if (p_settings.quality > RS::ENV_SSAO_QUALITY_VERY_LOW) {
-				blur_pipeline = SSAO_BLUR_PASS_SMART;
-				if (pass < blur_passes - 2) {
-					blur_pipeline = SSAO_BLUR_PASS_WIDE;
-				} else {
-					blur_pipeline = SSAO_BLUR_PASS_SMART;
-				}
-			}
-
-			for (int i = 0; i < 4; i++) {
-				if ((p_settings.quality == RS::ENV_SSAO_QUALITY_VERY_LOW) && ((i == 1) || (i == 2))) {
-					continue;
-				}
-
-				RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[blur_pipeline]);
-				if (pass % 2 == 0) {
-					if (p_settings.quality == RS::ENV_SSAO_QUALITY_VERY_LOW) {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ao_slices[i]), 0);
-					} else {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture_and_sampler(p_ao_slices[i], ss_effects.mirror_sampler), 0);
-					}
-
-					RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_ao_pong_slices[i]), 1);
-				} else {
-					if (p_settings.quality == RS::ENV_SSAO_QUALITY_VERY_LOW) {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ao_pong_slices[i]), 0);
-					} else {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture_and_sampler(p_ao_pong_slices[i], ss_effects.mirror_sampler), 0);
-					}
-					RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_ao_slices[i]), 1);
-				}
-				RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssao.blur_push_constant, sizeof(SSAOBlurPushConstant));
-
-				Size2i size(p_settings.full_screen_size.x >> (p_settings.half_size ? 2 : 1), p_settings.full_screen_size.y >> (p_settings.half_size ? 2 : 1));
-				RD::get_singleton()->compute_list_dispatch_threads(compute_list, size.x, size.y, 1);
-			}
-
-			if (p_settings.quality > RS::ENV_SSAO_QUALITY_VERY_LOW) {
-				RD::get_singleton()->compute_list_add_barrier(compute_list);
-			}
-		}
-		RD::get_singleton()->draw_command_end_label(); // Blur
-	}
-
-	/* FOURTH PASS */
-	// Interleave buffers
-	// back to full size
-	{
-		RD::get_singleton()->draw_command_begin_label("Interleave Buffers");
-		ssao.interleave_push_constant.inv_sharpness = 1.0 - p_settings.sharpness;
-		ssao.interleave_push_constant.pixel_size[0] = 1.0 / p_settings.full_screen_size.x;
-		ssao.interleave_push_constant.pixel_size[1] = 1.0 / p_settings.full_screen_size.y;
-		ssao.interleave_push_constant.size_modifier = uint32_t(p_settings.half_size ? 4 : 2);
-
-		int interleave_pipeline = SSAO_INTERLEAVE_HALF;
-		if (p_settings.quality == RS::ENV_SSAO_QUALITY_LOW) {
-			interleave_pipeline = SSAO_INTERLEAVE;
-		} else if (p_settings.quality >= RS::ENV_SSAO_QUALITY_MEDIUM) {
-			interleave_pipeline = SSAO_INTERLEAVE_SMART;
-		}
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssao.pipelines[interleave_pipeline]);
-
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_upscale_buffer), 0);
-		if (p_settings.quality > RS::ENV_SSAO_QUALITY_VERY_LOW && p_settings.blur_passes % 2 == 0) {
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ao), 1);
-		} else {
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ao_pong), 1);
-		}
-
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssao.interleave_push_constant, sizeof(SSAOInterleavePushConstant));
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.full_screen_size.x, p_settings.full_screen_size.y, 1);
-		RD::get_singleton()->compute_list_add_barrier(compute_list);
-		RD::get_singleton()->draw_command_end_label(); // Interleave
-	}
-	RD::get_singleton()->draw_command_end_label(); //SSAO
-	RD::get_singleton()->compute_list_end(RD::BARRIER_MASK_NO_BARRIER); //wait for upcoming transfer
-
-	int zero[1] = { 0 };
-	RD::get_singleton()->buffer_update(ssao.importance_map_load_counter, 0, sizeof(uint32_t), &zero, 0); //no barrier
-}
-
-void EffectsRD::gather_ssil(RD::ComputeListID p_compute_list, const Vector<RID> p_ssil_slices, const Vector<RID> p_edges_slices, const SSILSettings &p_settings, bool p_adaptive_base_pass, RID p_gather_uniform_set, RID p_importance_map_uniform_set, RID p_projection_uniform_set) {
-	RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, p_gather_uniform_set, 0);
-	if ((p_settings.quality == RS::ENV_SSIL_QUALITY_ULTRA) && !p_adaptive_base_pass) {
-		RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, p_importance_map_uniform_set, 1);
-	}
-	RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, p_projection_uniform_set, 3);
-
-	for (int i = 0; i < 4; i++) {
-		if ((p_settings.quality == RS::ENV_SSIL_QUALITY_VERY_LOW) && ((i == 1) || (i == 2))) {
-			continue;
-		}
-
-		ssil.gather_push_constant.pass_coord_offset[0] = i % 2;
-		ssil.gather_push_constant.pass_coord_offset[1] = i / 2;
-		ssil.gather_push_constant.pass_uv_offset[0] = ((i % 2) - 0.0) / p_settings.full_screen_size.x;
-		ssil.gather_push_constant.pass_uv_offset[1] = ((i / 2) - 0.0) / p_settings.full_screen_size.y;
-		ssil.gather_push_constant.pass = i;
-		RD::get_singleton()->compute_list_bind_uniform_set(p_compute_list, _get_compute_uniform_set_from_image_pair(p_ssil_slices[i], p_edges_slices[i]), 2);
-		RD::get_singleton()->compute_list_set_push_constant(p_compute_list, &ssil.gather_push_constant, sizeof(SSILGatherPushConstant));
-
-		Size2i size = Size2i(p_settings.full_screen_size.x >> (p_settings.half_size ? 2 : 1), p_settings.full_screen_size.y >> (p_settings.half_size ? 2 : 1));
-
-		RD::get_singleton()->compute_list_dispatch_threads(p_compute_list, size.x, size.y, 1);
-	}
-	RD::get_singleton()->compute_list_add_barrier(p_compute_list);
-}
-
-void EffectsRD::screen_space_indirect_lighting(RID p_diffuse, RID p_destination, RID p_normal_buffer, RID p_depth_mipmaps_texture, RID p_ssil, const Vector<RID> p_ssil_slices, RID p_ssil_pong, const Vector<RID> p_ssil_pong_slices, RID p_importance_map, RID p_importance_map_pong, RID p_edges, const Vector<RID> p_edges_slices, const CameraMatrix &p_projection, const CameraMatrix &p_last_projection, const SSILSettings &p_settings, bool p_invalidate_uniform_sets, RID &r_gather_uniform_set, RID &r_importance_map_uniform_set, RID &r_projection_uniform_set) {
-	RD::get_singleton()->draw_command_begin_label("Process Screen Space Indirect Lighting");
-	//Store projection info before starting the compute list
-	SSILProjectionUniforms projection_uniforms;
-	store_camera(p_last_projection, projection_uniforms.inv_last_frame_projection_matrix);
-
-	RD::get_singleton()->buffer_update(ssil.projection_uniform_buffer, 0, sizeof(SSILProjectionUniforms), &projection_uniforms);
-
-	memset(&ssil.gather_push_constant, 0, sizeof(SSILGatherPushConstant));
-
-	RD::ComputeListID compute_list = RD::get_singleton()->compute_list_begin();
-	{
-		RD::get_singleton()->draw_command_begin_label("Gather Samples");
-		ssil.gather_push_constant.screen_size[0] = p_settings.full_screen_size.x;
-		ssil.gather_push_constant.screen_size[1] = p_settings.full_screen_size.y;
-
-		ssil.gather_push_constant.half_screen_pixel_size[0] = 1.0 / p_settings.half_screen_size.x;
-		ssil.gather_push_constant.half_screen_pixel_size[1] = 1.0 / p_settings.half_screen_size.y;
-		float tan_half_fov_x = 1.0 / p_projection.matrix[0][0];
-		float tan_half_fov_y = 1.0 / p_projection.matrix[1][1];
-		ssil.gather_push_constant.NDC_to_view_mul[0] = tan_half_fov_x * 2.0;
-		ssil.gather_push_constant.NDC_to_view_mul[1] = tan_half_fov_y * -2.0;
-		ssil.gather_push_constant.NDC_to_view_add[0] = tan_half_fov_x * -1.0;
-		ssil.gather_push_constant.NDC_to_view_add[1] = tan_half_fov_y;
-		ssil.gather_push_constant.z_near = p_projection.get_z_near();
-		ssil.gather_push_constant.z_far = p_projection.get_z_far();
-		ssil.gather_push_constant.is_orthogonal = p_projection.is_orthogonal();
-
-		ssil.gather_push_constant.half_screen_pixel_size_x025[0] = ssil.gather_push_constant.half_screen_pixel_size[0] * 0.25;
-		ssil.gather_push_constant.half_screen_pixel_size_x025[1] = ssil.gather_push_constant.half_screen_pixel_size[1] * 0.25;
-
-		ssil.gather_push_constant.radius = p_settings.radius;
-		float radius_near_limit = (p_settings.radius * 1.2f);
-		if (p_settings.quality <= RS::ENV_SSIL_QUALITY_LOW) {
-			radius_near_limit *= 1.50f;
-
-			if (p_settings.quality == RS::ENV_SSIL_QUALITY_VERY_LOW) {
-				ssil.gather_push_constant.radius *= 0.8f;
-			}
-		}
-		radius_near_limit /= tan_half_fov_y;
-		ssil.gather_push_constant.intensity = p_settings.intensity * Math_PI;
-		ssil.gather_push_constant.fade_out_mul = -1.0 / (p_settings.fadeout_to - p_settings.fadeout_from);
-		ssil.gather_push_constant.fade_out_add = p_settings.fadeout_from / (p_settings.fadeout_to - p_settings.fadeout_from) + 1.0;
-		ssil.gather_push_constant.inv_radius_near_limit = 1.0f / radius_near_limit;
-		ssil.gather_push_constant.neg_inv_radius = -1.0 / ssil.gather_push_constant.radius;
-		ssil.gather_push_constant.normal_rejection_amount = p_settings.normal_rejection;
-
-		ssil.gather_push_constant.load_counter_avg_div = 9.0 / float((p_settings.quarter_screen_size.x) * (p_settings.quarter_screen_size.y) * 255);
-		ssil.gather_push_constant.adaptive_sample_limit = p_settings.adaptive_target;
-
-		ssil.gather_push_constant.quality = MAX(0, p_settings.quality - 1);
-		ssil.gather_push_constant.size_multiplier = p_settings.half_size ? 2 : 1;
-
-		if (p_invalidate_uniform_sets) {
-			Vector<RD::Uniform> uniforms;
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-				u.binding = 0;
-				u.append_id(default_mipmap_sampler);
-				u.append_id(p_diffuse);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_UNIFORM_BUFFER;
-				u.binding = 1;
-				u.append_id(ssil.projection_uniform_buffer);
-				uniforms.push_back(u);
-			}
-			r_projection_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssil.gather_shader.version_get_shader(ssil.gather_shader_version, 0), 3);
-		}
-
-		if (p_invalidate_uniform_sets) {
-			Vector<RD::Uniform> uniforms;
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-				u.binding = 0;
-				u.append_id(default_sampler);
-				u.append_id(p_depth_mipmaps_texture);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-				u.binding = 1;
-				u.append_id(p_normal_buffer);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_UNIFORM_BUFFER;
-				u.binding = 2;
-				u.append_id(ss_effects.gather_constants_buffer);
-				uniforms.push_back(u);
-			}
-			r_gather_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssil.gather_shader.version_get_shader(ssil.gather_shader_version, 0), 0);
-		}
-
-		if (p_invalidate_uniform_sets) {
-			Vector<RD::Uniform> uniforms;
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_IMAGE;
-				u.binding = 0;
-				u.append_id(p_ssil_pong);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_SAMPLER_WITH_TEXTURE;
-				u.binding = 1;
-				u.append_id(default_sampler);
-				u.append_id(p_importance_map);
-				uniforms.push_back(u);
-			}
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-				u.binding = 2;
-				u.append_id(ssil.importance_map_load_counter);
-				uniforms.push_back(u);
-			}
-			r_importance_map_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssil.gather_shader.version_get_shader(ssil.gather_shader_version, 2), 1);
-		}
-
-		if (p_settings.quality == RS::ENV_SSIL_QUALITY_ULTRA) {
-			RD::get_singleton()->draw_command_begin_label("Generate Importance Map");
-			ssil.importance_map_push_constant.half_screen_pixel_size[0] = 1.0 / p_settings.half_screen_size.x;
-			ssil.importance_map_push_constant.half_screen_pixel_size[1] = 1.0 / p_settings.half_screen_size.y;
-			ssil.importance_map_push_constant.intensity = p_settings.intensity * Math_PI;
-			//base pass
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[SSIL_GATHER_BASE]);
-			gather_ssil(compute_list, p_ssil_pong_slices, p_edges_slices, p_settings, true, r_gather_uniform_set, r_importance_map_uniform_set, r_projection_uniform_set);
-			//generate importance map
-
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[SSIL_GENERATE_IMPORTANCE_MAP]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ssil_pong), 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_importance_map), 1);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssil.importance_map_push_constant, sizeof(SSILImportanceMapPushConstant));
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.quarter_screen_size.x, p_settings.quarter_screen_size.y, 1);
-			RD::get_singleton()->compute_list_add_barrier(compute_list);
-			// process Importance Map A
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[SSIL_PROCESS_IMPORTANCE_MAPA]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_importance_map), 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_importance_map_pong), 1);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssil.importance_map_push_constant, sizeof(SSILImportanceMapPushConstant));
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.quarter_screen_size.x, p_settings.quarter_screen_size.y, 1);
-			RD::get_singleton()->compute_list_add_barrier(compute_list);
-			// process Importance Map B
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[SSIL_PROCESS_IMPORTANCE_MAPB]);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_importance_map_pong), 0);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_importance_map), 1);
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, ssil.counter_uniform_set, 2);
-			RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssil.importance_map_push_constant, sizeof(SSILImportanceMapPushConstant));
-			RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.quarter_screen_size.x, p_settings.quarter_screen_size.y, 1);
-			RD::get_singleton()->compute_list_add_barrier(compute_list);
-
-			RD::get_singleton()->draw_command_end_label(); // Importance Map
-
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[SSIL_GATHER_ADAPTIVE]);
-		} else {
-			RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[SSIL_GATHER]);
-		}
-
-		gather_ssil(compute_list, p_ssil_slices, p_edges_slices, p_settings, false, r_gather_uniform_set, r_importance_map_uniform_set, r_projection_uniform_set);
-		RD::get_singleton()->draw_command_end_label(); //Gather
-	}
-
-	{
-		RD::get_singleton()->draw_command_begin_label("Edge Aware Blur");
-		ssil.blur_push_constant.edge_sharpness = 1.0 - p_settings.sharpness;
-		ssil.blur_push_constant.half_screen_pixel_size[0] = 1.0 / p_settings.half_screen_size.x;
-		ssil.blur_push_constant.half_screen_pixel_size[1] = 1.0 / p_settings.half_screen_size.y;
-
-		int blur_passes = p_settings.quality > RS::ENV_SSIL_QUALITY_VERY_LOW ? p_settings.blur_passes : 1;
-
-		for (int pass = 0; pass < blur_passes; pass++) {
-			int blur_pipeline = SSIL_BLUR_PASS;
-			if (p_settings.quality > RS::ENV_SSIL_QUALITY_VERY_LOW) {
-				blur_pipeline = SSIL_BLUR_PASS_SMART;
-				if (pass < blur_passes - 2) {
-					blur_pipeline = SSIL_BLUR_PASS_WIDE;
-				}
-			}
-
-			for (int i = 0; i < 4; i++) {
-				if ((p_settings.quality == RS::ENV_SSIL_QUALITY_VERY_LOW) && ((i == 1) || (i == 2))) {
-					continue;
-				}
-
-				RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[blur_pipeline]);
-				if (pass % 2 == 0) {
-					if (p_settings.quality == RS::ENV_SSIL_QUALITY_VERY_LOW) {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ssil_slices[i]), 0);
-					} else {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture_and_sampler(p_ssil_slices[i], ss_effects.mirror_sampler), 0);
-					}
-					RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_ssil_pong_slices[i]), 1);
-				} else {
-					if (p_settings.quality == RS::ENV_SSIL_QUALITY_VERY_LOW) {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ssil_pong_slices[i]), 0);
-					} else {
-						RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture_and_sampler(p_ssil_pong_slices[i], ss_effects.mirror_sampler), 0);
-					}
-					RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_ssil_slices[i]), 1);
-				}
-				RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_edges_slices[i]), 2);
-
-				RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssil.blur_push_constant, sizeof(SSILBlurPushConstant));
-
-				int x_groups = (p_settings.full_screen_size.x >> (p_settings.half_size ? 2 : 1));
-				int y_groups = (p_settings.full_screen_size.y >> (p_settings.half_size ? 2 : 1));
-
-				RD::get_singleton()->compute_list_dispatch_threads(compute_list, x_groups, y_groups, 1);
-				if (p_settings.quality > RS::ENV_SSIL_QUALITY_VERY_LOW) {
-					RD::get_singleton()->compute_list_add_barrier(compute_list);
-				}
-			}
-		}
-
-		RD::get_singleton()->draw_command_end_label(); // Blur
-	}
-
-	{
-		RD::get_singleton()->draw_command_begin_label("Interleave Buffers");
-		ssil.interleave_push_constant.inv_sharpness = 1.0 - p_settings.sharpness;
-		ssil.interleave_push_constant.pixel_size[0] = 1.0 / p_settings.full_screen_size.x;
-		ssil.interleave_push_constant.pixel_size[1] = 1.0 / p_settings.full_screen_size.y;
-		ssil.interleave_push_constant.size_modifier = uint32_t(p_settings.half_size ? 4 : 2);
-
-		int interleave_pipeline = SSIL_INTERLEAVE_HALF;
-		if (p_settings.quality == RS::ENV_SSIL_QUALITY_LOW) {
-			interleave_pipeline = SSIL_INTERLEAVE;
-		} else if (p_settings.quality >= RS::ENV_SSIL_QUALITY_MEDIUM) {
-			interleave_pipeline = SSIL_INTERLEAVE_SMART;
-		}
-
-		RD::get_singleton()->compute_list_bind_compute_pipeline(compute_list, ssil.pipelines[interleave_pipeline]);
-
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_destination), 0);
-
-		if (p_settings.quality > RS::ENV_SSIL_QUALITY_VERY_LOW && p_settings.blur_passes % 2 == 0) {
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ssil), 1);
-		} else {
-			RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_compute_uniform_set_from_texture(p_ssil_pong), 1);
-		}
-		RD::get_singleton()->compute_list_bind_uniform_set(compute_list, _get_uniform_set_from_image(p_edges), 2);
-
-		RD::get_singleton()->compute_list_set_push_constant(compute_list, &ssil.interleave_push_constant, sizeof(SSILInterleavePushConstant));
-
-		RD::get_singleton()->compute_list_dispatch_threads(compute_list, p_settings.full_screen_size.x, p_settings.full_screen_size.y, 1);
-		RD::get_singleton()->compute_list_add_barrier(compute_list);
-		RD::get_singleton()->draw_command_end_label(); // Interleave
-	}
-
-	RD::get_singleton()->draw_command_end_label(); // SSIL
-
-	RD::get_singleton()->compute_list_end(RD::BARRIER_MASK_NO_BARRIER);
-
-	int zero[1] = { 0 };
-	RD::get_singleton()->buffer_update(ssil.importance_map_load_counter, 0, sizeof(uint32_t), &zero, 0); //no barrier
 }
 
 void EffectsRD::roughness_limit(RID p_source_normal, RID p_roughness, const Size2i &p_size, float p_curve) {
@@ -1358,154 +434,6 @@ EffectsRD::EffectsRD(bool p_prefer_raster_effects) {
 	}
 
 	if (!prefer_raster_effects) {
-		{
-			// Initialize depth buffer for screen space effects
-			Vector<String> downsampler_modes;
-			downsampler_modes.push_back("\n");
-			downsampler_modes.push_back("\n#define USE_HALF_SIZE\n");
-			downsampler_modes.push_back("\n#define GENERATE_MIPS\n");
-			downsampler_modes.push_back("\n#define GENERATE_MIPS\n#define USE_HALF_SIZE\n");
-			downsampler_modes.push_back("\n#define USE_HALF_BUFFERS\n");
-			downsampler_modes.push_back("\n#define USE_HALF_BUFFERS\n#define USE_HALF_SIZE\n");
-			downsampler_modes.push_back("\n#define GENERATE_MIPS\n#define GENERATE_FULL_MIPS");
-
-			ss_effects.downsample_shader.initialize(downsampler_modes);
-
-			ss_effects.downsample_shader_version = ss_effects.downsample_shader.version_create();
-
-			for (int i = 0; i < SS_EFFECTS_MAX; i++) {
-				ss_effects.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ss_effects.downsample_shader.version_get_shader(ss_effects.downsample_shader_version, i));
-			}
-
-			ss_effects.gather_constants_buffer = RD::get_singleton()->uniform_buffer_create(sizeof(SSEffectsGatherConstants));
-			SSEffectsGatherConstants gather_constants;
-
-			const int sub_pass_count = 5;
-			for (int pass = 0; pass < 4; pass++) {
-				for (int subPass = 0; subPass < sub_pass_count; subPass++) {
-					int a = pass;
-					int b = subPass;
-
-					int spmap[5]{ 0, 1, 4, 3, 2 };
-					b = spmap[subPass];
-
-					float ca, sa;
-					float angle0 = (float(a) + float(b) / float(sub_pass_count)) * Math_PI * 0.5f;
-
-					ca = Math::cos(angle0);
-					sa = Math::sin(angle0);
-
-					float scale = 1.0f + (a - 1.5f + (b - (sub_pass_count - 1.0f) * 0.5f) / float(sub_pass_count)) * 0.07f;
-
-					gather_constants.rotation_matrices[pass * 20 + subPass * 4 + 0] = scale * ca;
-					gather_constants.rotation_matrices[pass * 20 + subPass * 4 + 1] = scale * -sa;
-					gather_constants.rotation_matrices[pass * 20 + subPass * 4 + 2] = -scale * sa;
-					gather_constants.rotation_matrices[pass * 20 + subPass * 4 + 3] = -scale * ca;
-				}
-			}
-
-			RD::get_singleton()->buffer_update(ss_effects.gather_constants_buffer, 0, sizeof(SSEffectsGatherConstants), &gather_constants);
-		}
-
-		{
-			// Initialize ssao
-
-			RD::SamplerState sampler;
-			sampler.mag_filter = RD::SAMPLER_FILTER_NEAREST;
-			sampler.min_filter = RD::SAMPLER_FILTER_NEAREST;
-			sampler.mip_filter = RD::SAMPLER_FILTER_NEAREST;
-			sampler.repeat_u = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-			sampler.repeat_v = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-			sampler.repeat_w = RD::SAMPLER_REPEAT_MODE_MIRRORED_REPEAT;
-			sampler.max_lod = 4;
-
-			ss_effects.mirror_sampler = RD::get_singleton()->sampler_create(sampler);
-
-			uint32_t pipeline = 0;
-			{
-				Vector<String> ssao_modes;
-
-				ssao_modes.push_back("\n");
-				ssao_modes.push_back("\n#define SSAO_BASE\n");
-				ssao_modes.push_back("\n#define ADAPTIVE\n");
-
-				ssao.gather_shader.initialize(ssao_modes);
-
-				ssao.gather_shader_version = ssao.gather_shader.version_create();
-
-				for (int i = 0; i <= SSAO_GATHER_ADAPTIVE; i++) {
-					ssao.pipelines[pipeline] = RD::get_singleton()->compute_pipeline_create(ssao.gather_shader.version_get_shader(ssao.gather_shader_version, i));
-					pipeline++;
-				}
-			}
-			{
-				Vector<String> ssao_modes;
-				ssao_modes.push_back("\n#define GENERATE_MAP\n");
-				ssao_modes.push_back("\n#define PROCESS_MAPA\n");
-				ssao_modes.push_back("\n#define PROCESS_MAPB\n");
-
-				ssao.importance_map_shader.initialize(ssao_modes);
-
-				ssao.importance_map_shader_version = ssao.importance_map_shader.version_create();
-
-				for (int i = SSAO_GENERATE_IMPORTANCE_MAP; i <= SSAO_PROCESS_IMPORTANCE_MAPB; i++) {
-					ssao.pipelines[pipeline] = RD::get_singleton()->compute_pipeline_create(ssao.importance_map_shader.version_get_shader(ssao.importance_map_shader_version, i - SSAO_GENERATE_IMPORTANCE_MAP));
-
-					pipeline++;
-				}
-				ssao.importance_map_load_counter = RD::get_singleton()->storage_buffer_create(sizeof(uint32_t));
-				int zero[1] = { 0 };
-				RD::get_singleton()->buffer_update(ssao.importance_map_load_counter, 0, sizeof(uint32_t), &zero);
-				RD::get_singleton()->set_resource_name(ssao.importance_map_load_counter, "Importance Map Load Counter");
-
-				Vector<RD::Uniform> uniforms;
-				{
-					RD::Uniform u;
-					u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-					u.binding = 0;
-					u.append_id(ssao.importance_map_load_counter);
-					uniforms.push_back(u);
-				}
-				ssao.counter_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssao.importance_map_shader.version_get_shader(ssao.importance_map_shader_version, 2), 2);
-				RD::get_singleton()->set_resource_name(ssao.counter_uniform_set, "Load Counter Uniform Set");
-			}
-			{
-				Vector<String> ssao_modes;
-				ssao_modes.push_back("\n#define MODE_NON_SMART\n");
-				ssao_modes.push_back("\n#define MODE_SMART\n");
-				ssao_modes.push_back("\n#define MODE_WIDE\n");
-
-				ssao.blur_shader.initialize(ssao_modes);
-
-				ssao.blur_shader_version = ssao.blur_shader.version_create();
-
-				for (int i = SSAO_BLUR_PASS; i <= SSAO_BLUR_PASS_WIDE; i++) {
-					ssao.pipelines[pipeline] = RD::get_singleton()->compute_pipeline_create(ssao.blur_shader.version_get_shader(ssao.blur_shader_version, i - SSAO_BLUR_PASS));
-
-					pipeline++;
-				}
-			}
-			{
-				Vector<String> ssao_modes;
-				ssao_modes.push_back("\n#define MODE_NON_SMART\n");
-				ssao_modes.push_back("\n#define MODE_SMART\n");
-				ssao_modes.push_back("\n#define MODE_HALF\n");
-
-				ssao.interleave_shader.initialize(ssao_modes);
-
-				ssao.interleave_shader_version = ssao.interleave_shader.version_create();
-				for (int i = SSAO_INTERLEAVE; i <= SSAO_INTERLEAVE_HALF; i++) {
-					ssao.pipelines[pipeline] = RD::get_singleton()->compute_pipeline_create(ssao.interleave_shader.version_get_shader(ssao.interleave_shader_version, i - SSAO_INTERLEAVE));
-					RD::get_singleton()->set_resource_name(ssao.pipelines[pipeline], "Interleave Pipeline " + itos(i));
-					pipeline++;
-				}
-			}
-
-			ERR_FAIL_COND(pipeline != SSAO_MAX);
-		}
-	}
-
-	if (!prefer_raster_effects) {
 		// Initialize roughness limiter
 		Vector<String> shader_modes;
 		shader_modes.push_back("");
@@ -1518,81 +446,6 @@ EffectsRD::EffectsRD(bool p_prefer_raster_effects) {
 	}
 
 	if (!prefer_raster_effects) {
-		Vector<String> specular_modes;
-		specular_modes.push_back("\n#define MODE_MERGE\n");
-		specular_modes.push_back("\n#define MODE_MERGE\n#define MODE_SSR\n");
-		specular_modes.push_back("\n");
-		specular_modes.push_back("\n#define MODE_SSR\n");
-
-		specular_merge.shader.initialize(specular_modes);
-
-		specular_merge.shader_version = specular_merge.shader.version_create();
-
-		//use additive
-
-		RD::PipelineColorBlendState::Attachment ba;
-		ba.enable_blend = true;
-		ba.src_color_blend_factor = RD::BLEND_FACTOR_ONE;
-		ba.dst_color_blend_factor = RD::BLEND_FACTOR_ONE;
-		ba.src_alpha_blend_factor = RD::BLEND_FACTOR_ONE;
-		ba.dst_alpha_blend_factor = RD::BLEND_FACTOR_ONE;
-		ba.color_blend_op = RD::BLEND_OP_ADD;
-		ba.alpha_blend_op = RD::BLEND_OP_ADD;
-
-		RD::PipelineColorBlendState blend_additive;
-		blend_additive.attachments.push_back(ba);
-
-		for (int i = 0; i < SPECULAR_MERGE_MAX; i++) {
-			RD::PipelineColorBlendState blend_state;
-			if (i == SPECULAR_MERGE_ADDITIVE_ADD || i == SPECULAR_MERGE_ADDITIVE_SSR) {
-				blend_state = blend_additive;
-			} else {
-				blend_state = RD::PipelineColorBlendState::create_disabled();
-			}
-			specular_merge.pipelines[i].setup(specular_merge.shader.version_get_shader(specular_merge.shader_version, i), RD::RENDER_PRIMITIVE_TRIANGLES, RD::PipelineRasterizationState(), RD::PipelineMultisampleState(), RD::PipelineDepthStencilState(), blend_state, 0);
-		}
-	}
-
-	if (!prefer_raster_effects) {
-		{
-			Vector<String> ssr_modes;
-			ssr_modes.push_back("\n");
-			ssr_modes.push_back("\n#define MODE_ROUGH\n");
-
-			ssr.shader.initialize(ssr_modes);
-
-			ssr.shader_version = ssr.shader.version_create();
-
-			for (int i = 0; i < SCREEN_SPACE_REFLECTION_MAX; i++) {
-				ssr.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ssr.shader.version_get_shader(ssr.shader_version, i));
-			}
-		}
-
-		{
-			Vector<String> ssr_filter_modes;
-			ssr_filter_modes.push_back("\n");
-			ssr_filter_modes.push_back("\n#define VERTICAL_PASS\n");
-
-			ssr_filter.shader.initialize(ssr_filter_modes);
-
-			ssr_filter.shader_version = ssr_filter.shader.version_create();
-
-			for (int i = 0; i < SCREEN_SPACE_REFLECTION_FILTER_MAX; i++) {
-				ssr_filter.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ssr_filter.shader.version_get_shader(ssr_filter.shader_version, i));
-			}
-		}
-
-		{
-			Vector<String> ssr_scale_modes;
-			ssr_scale_modes.push_back("\n");
-
-			ssr_scale.shader.initialize(ssr_scale_modes);
-
-			ssr_scale.shader_version = ssr_scale.shader.version_create();
-
-			ssr_scale.pipeline = RD::get_singleton()->compute_pipeline_create(ssr_scale.shader.version_get_shader(ssr_scale.shader_version, 0));
-		}
-
 		{
 			Vector<String> sss_modes;
 			sss_modes.push_back("\n#define USE_11_SAMPLES\n");
@@ -1605,79 +458,6 @@ EffectsRD::EffectsRD(bool p_prefer_raster_effects) {
 
 			for (int i = 0; i < sss_modes.size(); i++) {
 				sss.pipelines[i] = RD::get_singleton()->compute_pipeline_create(sss.shader.version_get_shader(sss.shader_version, i));
-			}
-		}
-
-		{
-			Vector<String> ssil_modes;
-			ssil_modes.push_back("\n");
-			ssil_modes.push_back("\n#define SSIL_BASE\n");
-			ssil_modes.push_back("\n#define ADAPTIVE\n");
-
-			ssil.gather_shader.initialize(ssil_modes);
-
-			ssil.gather_shader_version = ssil.gather_shader.version_create();
-
-			for (int i = SSIL_GATHER; i <= SSIL_GATHER_ADAPTIVE; i++) {
-				ssil.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ssil.gather_shader.version_get_shader(ssil.gather_shader_version, i));
-			}
-			ssil.projection_uniform_buffer = RD::get_singleton()->uniform_buffer_create(sizeof(SSILProjectionUniforms));
-		}
-
-		{
-			Vector<String> ssil_modes;
-			ssil_modes.push_back("\n#define GENERATE_MAP\n");
-			ssil_modes.push_back("\n#define PROCESS_MAPA\n");
-			ssil_modes.push_back("\n#define PROCESS_MAPB\n");
-
-			ssil.importance_map_shader.initialize(ssil_modes);
-
-			ssil.importance_map_shader_version = ssil.importance_map_shader.version_create();
-
-			for (int i = SSIL_GENERATE_IMPORTANCE_MAP; i <= SSIL_PROCESS_IMPORTANCE_MAPB; i++) {
-				ssil.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ssil.importance_map_shader.version_get_shader(ssil.importance_map_shader_version, i - SSIL_GENERATE_IMPORTANCE_MAP));
-			}
-			ssil.importance_map_load_counter = RD::get_singleton()->storage_buffer_create(sizeof(uint32_t));
-			int zero[1] = { 0 };
-			RD::get_singleton()->buffer_update(ssil.importance_map_load_counter, 0, sizeof(uint32_t), &zero);
-			RD::get_singleton()->set_resource_name(ssil.importance_map_load_counter, "Importance Map Load Counter");
-
-			Vector<RD::Uniform> uniforms;
-			{
-				RD::Uniform u;
-				u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-				u.binding = 0;
-				u.append_id(ssil.importance_map_load_counter);
-				uniforms.push_back(u);
-			}
-			ssil.counter_uniform_set = RD::get_singleton()->uniform_set_create(uniforms, ssil.importance_map_shader.version_get_shader(ssil.importance_map_shader_version, 2), 2);
-			RD::get_singleton()->set_resource_name(ssil.counter_uniform_set, "Load Counter Uniform Set");
-		}
-		{
-			Vector<String> ssil_modes;
-			ssil_modes.push_back("\n#define MODE_NON_SMART\n");
-			ssil_modes.push_back("\n#define MODE_SMART\n");
-			ssil_modes.push_back("\n#define MODE_WIDE\n");
-
-			ssil.blur_shader.initialize(ssil_modes);
-
-			ssil.blur_shader_version = ssil.blur_shader.version_create();
-			for (int i = SSIL_BLUR_PASS; i <= SSIL_BLUR_PASS_WIDE; i++) {
-				ssil.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ssil.blur_shader.version_get_shader(ssil.blur_shader_version, i - SSIL_BLUR_PASS));
-			}
-		}
-
-		{
-			Vector<String> ssil_modes;
-			ssil_modes.push_back("\n#define MODE_NON_SMART\n");
-			ssil_modes.push_back("\n#define MODE_SMART\n");
-			ssil_modes.push_back("\n#define MODE_HALF\n");
-
-			ssil.interleave_shader.initialize(ssil_modes);
-
-			ssil.interleave_shader_version = ssil.interleave_shader.version_create();
-			for (int i = SSIL_INTERLEAVE; i <= SSIL_INTERLEAVE_HALF; i++) {
-				ssil.pipelines[i] = RD::get_singleton()->compute_pipeline_create(ssil.interleave_shader.version_get_shader(ssil.interleave_shader_version, i - SSIL_INTERLEAVE));
 			}
 		}
 	}
@@ -1751,27 +531,8 @@ EffectsRD::~EffectsRD() {
 		luminance_reduce.shader.version_free(luminance_reduce.shader_version);
 	}
 	if (!prefer_raster_effects) {
-		specular_merge.shader.version_free(specular_merge.shader_version);
-		ss_effects.downsample_shader.version_free(ss_effects.downsample_shader_version);
-		ssao.blur_shader.version_free(ssao.blur_shader_version);
-		ssao.gather_shader.version_free(ssao.gather_shader_version);
-		ssao.interleave_shader.version_free(ssao.interleave_shader_version);
-		ssao.importance_map_shader.version_free(ssao.importance_map_shader_version);
-		ssil.blur_shader.version_free(ssil.blur_shader_version);
-		ssil.gather_shader.version_free(ssil.gather_shader_version);
-		ssil.interleave_shader.version_free(ssil.interleave_shader_version);
-		ssil.importance_map_shader.version_free(ssil.importance_map_shader_version);
 		roughness_limiter.shader.version_free(roughness_limiter.shader_version);
-		ssr.shader.version_free(ssr.shader_version);
-		ssr_filter.shader.version_free(ssr_filter.shader_version);
-		ssr_scale.shader.version_free(ssr_scale.shader_version);
 		sss.shader.version_free(sss.shader_version);
-
-		RD::get_singleton()->free(ss_effects.mirror_sampler);
-		RD::get_singleton()->free(ss_effects.gather_constants_buffer);
-		RD::get_singleton()->free(ssao.importance_map_load_counter);
-		RD::get_singleton()->free(ssil.importance_map_load_counter);
-		RD::get_singleton()->free(ssil.projection_uniform_buffer);
 	}
 	sort.shader.version_free(sort.shader_version);
 }
