+++ servers/rendering/renderer_rd/shaders/effects/screen_space_reflection_scale.glsl
@@ -6,6 +6,11 @@
 
 layout(local_size_x = 8, local_size_y = 8, local_size_z = 1) in;
 
+/* Specialization Constants (Toggles) */
+
+layout(constant_id = 0) const bool sc_multiview = false;
+
+/* inputs */
 layout(set = 0, binding = 0) uniform sampler2D source_ssr;
 layout(set = 1, binding = 0) uniform sampler2D source_depth;
 layout(set = 1, binding = 1) uniform sampler2D source_normal;
@@ -28,7 +33,7 @@ void main() {
 	// Pixel being shaded
 	ivec2 ssC = ivec2(gl_GlobalInvocationID.xy);
 
-	if (any(greaterThanEqual(ssC, params.screen_size))) { //too large, do nothing
+	if (any(greaterThanEqual(ssC.xy, params.screen_size))) { //too large, do nothing
 		return;
 	}
 	//do not filter, SSR will generate arctifacts if this is done
@@ -57,13 +62,19 @@ void main() {
 			normal.xyz += nr.xyz * 2.0 - 1.0;
 			normal.w += nr.w;
 
-			d = d * 2.0 - 1.0;
-			if (params.orthogonal) {
-				d = ((d + (params.camera_z_far + params.camera_z_near) / (params.camera_z_far - params.camera_z_near)) * (params.camera_z_far - params.camera_z_near)) / 2.0;
+			if (sc_multiview) {
+				// we're doing a full unproject so we need the value as is.
+				depth += d;
 			} else {
-				d = 2.0 * params.camera_z_near * params.camera_z_far / (params.camera_z_far + params.camera_z_near - d * (params.camera_z_far - params.camera_z_near));
+				// unproject our Z value so we can use it directly.
+				d = d * 2.0 - 1.0;
+				if (params.orthogonal) {
+					d = ((d + (params.camera_z_far + params.camera_z_near) / (params.camera_z_far - params.camera_z_near)) * (params.camera_z_far - params.camera_z_near)) / 2.0;
+				} else {
+					d = 2.0 * params.camera_z_near * params.camera_z_far / (params.camera_z_far + params.camera_z_near - d * (params.camera_z_far - params.camera_z_near));
+				}
+				depth += -d;
 			}
-			depth += -d;
 		}
 
 		color /= 4.0;
@@ -71,17 +82,22 @@ void main() {
 		normal.xyz = normalize(normal.xyz / 4.0) * 0.5 + 0.5;
 		normal.w /= 4.0;
 	} else {
-		color = texelFetch(source_ssr, ssC << 1, 0);
-		depth = texelFetch(source_depth, ssC << 1, 0).r;
-		normal = texelFetch(source_normal, ssC << 1, 0);
+		ivec2 ofs = ssC << 1;
 
-		depth = depth * 2.0 - 1.0;
-		if (params.orthogonal) {
-			depth = ((depth + (params.camera_z_far + params.camera_z_near) / (params.camera_z_far - params.camera_z_near)) * (params.camera_z_far - params.camera_z_near)) / 2.0;
-		} else {
-			depth = 2.0 * params.camera_z_near * params.camera_z_far / (params.camera_z_far + params.camera_z_near - depth * (params.camera_z_far - params.camera_z_near));
+		color = texelFetch(source_ssr, ofs, 0);
+		depth = texelFetch(source_depth, ofs, 0).r;
+		normal = texelFetch(source_normal, ofs, 0);
+
+		if (!sc_multiview) {
+			// unproject our Z value so we can use it directly.
+			depth = depth * 2.0 - 1.0;
+			if (params.orthogonal) {
+				depth = ((depth + (params.camera_z_far + params.camera_z_near) / (params.camera_z_far - params.camera_z_near)) * (params.camera_z_far - params.camera_z_near)) / 2.0;
+			} else {
+				depth = 2.0 * params.camera_z_near * params.camera_z_far / (params.camera_z_far + params.camera_z_near - depth * (params.camera_z_far - params.camera_z_near));
+			}
+			depth = -depth;
 		}
-		depth = -depth;
 	}
 
 	imageStore(dest_ssr, ssC, color);
