+++ servers/rendering/renderer_rd/shaders/effects/ss_effects_downsample.glsl
