+++ servers/rendering/renderer_rd/shaders/effects/ssao.glsl
