+++ servers/rendering/renderer_rd/shaders/effects/ssil.glsl
