+++ servers/rendering/renderer_rd/shaders/effects/ssil_blur.glsl
